+++ apps/prairielearn/src/lib/features/index.ts
@@ -1,5 +1,10 @@
-import { FeatureManager, FeatureGrantType } from './manager';
+import { FeatureManager } from './manager';
+import { featuresMiddleware } from './middleware';
 
-const features = new FeatureManager(['manual-grading-rubrics']);
+const features = new FeatureManager([
+  'manual-grading-rubrics',
+  // Can only be applied to courses/institutions.
+  'process-questions-in-worker',
+]);
 
-export { features, FeatureGrantType };
+export { features, featuresMiddleware };