+++ apps/prairielearn/src/migrations/20221122001508_time_series__date__add_index.sql
@@ -1,1 +1,2 @@
+-- prairielearn:migrations NO TRANSACTION
 CREATE INDEX CONCURRENTLY time_series_date_idx ON time_series (date);