+++ packages/html-ejs/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @prairielearn/html-ejs
 
+## 1.1.4
+
+### Patch Changes
+
+- Updated dependencies [2c5504f1f]
+  - @prairielearn/html@3.0.2
+
+## 1.1.3
+
+### Patch Changes
+
+- 8fd47d928: Upgrade all dependencies
+- Updated dependencies [8fd47d928]
+  - @prairielearn/html@3.0.1
+
 ## 1.1.2
 
 ### Patch Changes