+++ packages/migrations/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @prairielearn/migrations
 
+## 2.0.1
+
+### Patch Changes
+
+- 8fd47d928: Upgrade all dependencies
+- Updated dependencies [8fd47d928]
+  - @prairielearn/named-locks@1.3.3
+  - @prairielearn/postgres@1.7.1
+  - @prairielearn/logger@1.0.1
+  - @prairielearn/error@1.0.2
+
+## 2.0.0
+
+### Major Changes
+
+- 765dc616e: SQL migrations are now run inside a transaction by default
+
+  If you need to disable this behavior, you can add an annotation comment to the top of the migration file:
+
+  ```sql
+  -- prairielearn:migrations NO TRANSACTION
+  CREATE INDEX CONCURRENTLY ...
+  ```
+
+  Note that migrations implemented as JavaScript are not run in transactions by default. Transactions should be manually used as appropriate.
+
 ## 1.2.2
 
 ### Patch Changes