+++ packages/workspace-utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @prairielearn/workspace-utils
 
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies [8fd47d928]
+  - @prairielearn/postgres@1.7.1
+
 ## 1.0.4
 
 ### Patch Changes